+++ lib/geordi/commands/chromedriver_update.rb
@@ -7,11 +7,11 @@ installed Chrome.
 
 Setting `auto_update_chromedriver` to `true` in your global Geordi config file 
 (`~/.config/geordi/global.yml`), will automatically update chromedriver before 
-cucumber tests, in case Chrome and chromedriver versions don't match
+cucumber tests if a newer chromedriver version is available.
 LONGDESC
 
 option :quiet_if_matching, type: :boolean, default: false,
-  desc: 'Suppress notification if chromedriver and chrome versions match'
+  desc: 'Suppress notification if chromedriver is already on the latest version'
 
 def chromedriver_update
   require 'geordi/chromedriver_updater'